Adhering to samples structure
