+++ samples/python/azureai-streaming/env.TEMPLATE
@@ -1,0 +1,6 @@
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__CLIENTID=
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__CLIENTSECRET=
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__TENANTID=
+
+AZURE_OPENAI_API_VERSION=
+AZURE_OPENAI_ENDPOINT=
