+++ samples/python/azureai-streaming/requirements.txt
@@ -1,0 +1,8 @@
+python-dotenv
+aiohttp
+openai
+azure-identity
+microsoft-agents-activity
+microsoft-agents-hosting-core
+microsoft-agents-hosting-aiohttp
+microsoft-agents-authentication-msal
