+++ samples/python/cards/src/main.py
@@ -2,7 +2,7 @@
 # Licensed under the MIT License.
 
 # enable logging for Microsoft Agents library
-# for more information, see README.md for empty-agent sample
+# for more information, see README.md for Quickstart Agent
 import logging
 ms_agents_logger = logging.getLogger("microsoft.agents")
 ms_agents_logger.addHandler(logging.StreamHandler())
