--- samples/python/empty-agent/env.TEMPLATE
@@ -1,3 +1,0 @@
-CONNECTIONS__SERVICE_CONNECTION__SETTINGS__CLIENTID=client-id
-CONNECTIONS__SERVICE_CONNECTION__SETTINGS__CLIENTSECRET=client-secret
-CONNECTIONS__SERVICE_CONNECTION__SETTINGS__TENANTID=tenant-id
