--- samples/python/empty-agent/src/agent.py
@@ -1,64 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-
-import os.path as path
-import re
-import sys
-import traceback
-from dotenv import load_dotenv
-
-from os import environ
-from microsoft.agents.hosting.aiohttp import CloudAdapter
-from microsoft.agents.hosting.core import (
-    Authorization,
-    AgentApplication,
-    TurnState,
-    TurnContext,
-    MemoryStorage,
-)
-from microsoft.agents.authentication.msal import MsalConnectionManager
-from microsoft.agents.activity import load_configuration_from_env
-
-load_dotenv()  # robrandao: todo
-agents_sdk_config = load_configuration_from_env(environ)
-
-STORAGE = MemoryStorage()
-CONNECTION_MANAGER = MsalConnectionManager(**agents_sdk_config)
-ADAPTER = CloudAdapter(connection_manager=CONNECTION_MANAGER)
-AUTHORIZATION = Authorization(STORAGE, CONNECTION_MANAGER, **agents_sdk_config)
-
-# robrandao: downloader?
-AGENT_APP = AgentApplication[TurnState](
-    storage=STORAGE, adapter=ADAPTER, authorization=AUTHORIZATION, **agents_sdk_config
-)
-
-
-@AGENT_APP.conversation_update("membersAdded")
-async def on_members_added(context: TurnContext, _state: TurnState):
-    await context.send_activity(
-        "Welcome to the empty agent! "
-        "This agent is designed to be a starting point for your own agent development."
-    )
-    return True
-
-
-@AGENT_APP.message(re.compile(r"^hello$"))
-async def on_hello(context: TurnContext, _state: TurnState):
-    await context.send_activity("Hello!")
-
-
-@AGENT_APP.activity("message")
-async def on_message(context: TurnContext, _state: TurnState):
-    await context.send_activity(f"you said: {context.activity.text}")
-
-
-@AGENT_APP.error
-async def on_error(context: TurnContext, error: Exception):
-    # This check writes out errors to console log .vs. app insights.
-    # NOTE: In production environment, you should consider logging this to Azure
-    #       application insights.
-    print(f"\n [on_turn_error] unhandled error: {error}", file=sys.stderr)
-    traceback.print_exc()
-
-    # Send a message to the user
-    await context.send_activity("The bot encountered an error or bug.")
