+++ samples/python/quickstart/env.TEMPLATE
@@ -1,0 +1,3 @@
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__CLIENTID=client-id
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__CLIENTSECRET=client-secret
+CONNECTIONS__SERVICE_CONNECTION__SETTINGS__TENANTID=tenant-id
