+++ samples/python/quickstart/requirements.txt
@@ -1,0 +1,6 @@
+python-dotenv
+aiohttp
+microsoft-agents-hosting-aiohttp
+microsoft-agents-hosting-core
+microsoft-agents-authentication-msal
+microsoft-agents-activity
