+++ samples/python/semantic-kernel-multiturn/requirements.txt
@@ -1,0 +1,7 @@
+pydantic
+semantic-kernel
+azure-identity
+microsoft-agents-authentication-msal
+microsoft-agents-hosting-core
+microsoft-agents-hosting-aiohttp
+microsoft-agents-activity
