+++ samples/python/semantic-kernel-multiturn/src/main.py
@@ -1,0 +1,17 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+# enable logging for Microsoft Agents library
+# for more information, see README.md for Quickstart Agent
+import logging
+ms_agents_logger = logging.getLogger("microsoft.agents")
+ms_agents_logger.addHandler(logging.StreamHandler())
+ms_agents_logger.setLevel(logging.INFO)
+
+from .app import AGENT_APP, CONNECTION_MANAGER
+from .start_server import start_server
+
+start_server(
+    agent_application=AGENT_APP,
+    auth_configuration=CONNECTION_MANAGER.get_default_connection_configuration(),
+)
