+++ samples/python/semantic-kernel-multiturn/src/plugins/weather_forecast.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class WeatherForecast(BaseModel):
+    date: str
+    temperatureC: int
+    temperatureF: int
