--- samples/python/weather-agent/requirements.txt
@@ -1,7 +1,0 @@
-pydantic
-semantic-kernel
-azure-identity
-microsoft-agents-authentication-msal
-microsoft-agents-hosting-core
-microsoft-agents-hosting-aiohttp
-microsoft-agents-activity
