--- samples/python/weather-agent/src/plugins/__init__.py
@@ -1,4 +1,0 @@
-from .adaptive_card_plugin import *
-from .date_time_plugin import *
-from .weather_forecast_plugin import *
-from .weather_forecast import *
