--- samples/python/weather-agent/src/plugins/weather_forecast.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-class WeatherForecast(BaseModel):
-    date: str
-    temperatureC: int
-    temperatureF: int
